chore: move test route logic to separate file
